apigateway/missions/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Amichay Oren", email = "amichay@nine30.com"},
]
dependencies = [
    "aiohttp==3.8.4",
    "aiosignal==1.3.1",
    "async-timeout==4.0.2",
    "attrs==23.1.0",
    "autoflake==2.2.0",
    "bandit==1.7.5",
    "blinker==1.6.2",
    "cachetools==5.3.0",
    "certifi==2023.5.7",
    "cffi==1.15.1",
    "charset-normalizer==3.1.0",
    "click==8.1.3",
    "cloudevents==1.9.0",
    "cryptography==40.0.2",
    "deprecation==2.1.0",
    "flask==2.3.2",
    "frozenlist==1.3.3",
    "functions-framework==3.4.0",
    "gitdb==4.0.10",
    "GitPython==3.1.31",
    "google-api-core==2.11.1",
    "google-auth==2.19.0",
    "google-cloud==0.34.0",
    "google-cloud-core==2.3.3",
    "google-cloud-firestore==2.11.1",
    "google-cloud-pubsub==2.17.1",
    "google-cloud-secret-manager==2.16.2",
    "google-crc32c==1.5.0",
    "googleapis-common-protos==1.59.1",
    "grpc-google-iam-v1==0.12.6",
    "grpcio==1.54.2",
    "grpcio-status==1.54.2",
    "gunicorn==20.1.0",
    "idna==3.4",
    "importlib-metadata==6.6.0",
    "inflection==0.5.1",
    "isort==5.12.0",
    "itsdangerous==2.1.2",
    "jaraco-classes==3.2.3",
    "jeepney==0.8.0",
    "Jinja2==3.1.2",
    "jsonschema==4.17.3",
    "keyring==23.13.1",
    "keyrings-google-artifactregistry-auth==1.1.2",
    "markdown-it-py==2.2.0",
    "MarkupSafe==2.1.2",
    "mdurl==0.1.2",
    "more-itertools==9.1.0",
    "multidict==6.0.4",
    "openai==0.27.8",
    "openapi==1.1.0",
    "packaging==23.1",
    "pbr==5.11.1",
    "pluggy==1.0.0",
    "proto-plus==1.22.2",
    "protobuf==4.23.2",
    "pyasn1==0.5.0",
    "pyasn1-modules==0.3.0",
    "pycparser==2.21",
    "pyflakes==3.0.1",
    "pygments==2.15.1",
    "pyrsistent==0.19.3",
    "PyYAML==6.0",
    "requests==2.31.0",
    "rich==13.3.5",
    "rsa==4.9",
    "SecretStorage==3.3.3",
    "setuptools==68.0.0",
    "six==1.16.0",
    "smmap==5.0.0",
    "stevedore==5.1.0",
    "tqdm==4.65.0",
    "urllib3==1.26.16",
    "watchdog==3.0.0",
    "Werkzeug==2.3.4",
    "yarl==1.9.2",
    "zipp==3.15.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

